+++ split.py
@@ -20,7 +20,7 @@ from segtypes.linker_entry import (
 from segtypes.segment import Segment
 from util import log, options, palettes, symbols, relocs
 
-VERSION = "0.16.3"
+VERSION = "0.16.4"
 
 parser = argparse.ArgumentParser(
     description="Split a rom given a rom, a config, and output directory"